STRING SORTING
This Java program sorts a list of user-entered names in alphabetical order. It begins by prompting the user for the total number of names to sort, then stores each entered name in an array. To achieve alphabetical sorting, the program uses a nested loop to compare each pair of names, swapping them if they are not in the correct order. This comparison and swapping ensure that names are sorted in ascending alphabetical order by the end of the process. Finally, the program outputs the names in their newly sorted order, displaying them one by one in alphabetical sequence.